Simplify calculation of intersection and improve documentation.
Because our choice of slopes guarantees a zero determinant for the quadratic
representing intersections of a line and circle, we can simplify the
calculation of the value `x` for which the line and circle intersect.

Furthermore, improving the documentation of this section to explain why
this is possible and to add subscripts to the algebraic explanation of
the quadratic solution (previously there were variables `x` and `cx`
and an instance of `cx^2` could be easily confused as looking like
the `x` was being squared.  Now, it is shown as `x`, `c_x` and the
expression `c_x^2`, to make the algebraic guide more clear.

This also removes the issue where slight floating point arithmetic
caclulations would lead to complex number results (with a miniscule
complex component, such as:
{(430.94494865162426+5.8882236608733186e-06j),
 (430.94494865162426-5.8882236608733186e-06j)}
Removing this issue allows us to drop the `.real` post processing
of the real valued (x,y) intersection point.